requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Local LLM (Ollama) - No OpenAI needed!
ollama==0.1.7
langchain==0.0.350
langchain-community==0.0.8

# Local Embeddings - Fixed versions for compatibility
sentence-transformers==2.3.1
transformers==4.36.0
huggingface_hub==0.20.3

# Vector Database
chromadb==0.4.18

# Authentication & JWT
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
pypdf2==3.0.1
python-docx==1.1.0

asyncpg==0.29.0
redis[hiredis]==5.0.1
python-dotenv==1.0.0


# Utils
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
tiktoken==0.5.2